✅TEST : Service 테스트 코드 분할
- ServiceSpy 테스트 코드 삭제
